+++ app/Http/Controllers/TagsController.php
@@ -102,6 +102,10 @@ class TagsController extends Controller
     }
 
     public function destroy(Tag $tag){
+        if($tag->posts->count() > 0){
+            session()->flash('error','Tag can\'t be deleted becuase it has some posts.');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success', 'Tag Deleted Successfully');
